Added fix for mssqlnative to install/install_functions.php

+++ root/install/install_functions.php
@@ -154,7 +154,7 @@ function b3p_drop_table($table)
 
 	$table_name = substr($table . '#', 6, -1);
 
-	if ($db->sql_layer != 'mssql' && $db->sql_layer != 'mssql_odbc')
+	if ($db->sql_layer != 'mssql' && $db->sql_layer != 'mssql_odbc' && $db->sql_layer != 'mssqlnative')
 	{
 		$sql = 'DROP TABLE IF EXISTS ' . $table_prefix . $table_name;
 		$result = $db->sql_query($sql);
